test_patient_gui.py
import tkinter as tk
from tkinter import ttk, filedialog, messagebox
import pytest
import matplotlib.pyplot as plt
import numpy as np
from PIL import Image, ImageTk
import tempfile


def test_select_cpap_file(monkeypatch):
    from patient_gui import select_cpap_file
    file_path = "path/to/file.txt"
    monkeypatch.setattr(filedialog,
                        "askopenfilename",
                        lambda *args,
                        **kwargs: file_path)
    monkeypatch.setattr(messagebox,
                        "showerror",
                        lambda *args,
                        **kwargs: None)
    selected_file = select_cpap_file()
    assert selected_file == file_path


@pytest.mark.parametrize("file_path, expected_keys", [
    (None, ["breath_rate_bpm", "apnea_count",
            "flow_rate", "filtered_flow_rate"]),
    ("invalid_file_path.txt", None),
])
def test_process_cpap_data(file_path, expected_keys, monkeypatch):
    from patient_gui import process_cpap_data

    def mock_process_cpap_data(file_path):
        if file_path == "invalid_file_path.txt":
            return None
        elif file_path is None:
            raise TypeError(("expected str, bytes or "
                             "os.PathLike object, not NoneType"))
        else:
            return {
                "breath_rate_bpm": 15,
                "apnea_count": 1,
                "flow_rate": np.array([[0, 1], [1, 2], [2, 3]]),
                "filtered_flow_rate": np.array([[0, 1], [1, 2], [2, 3]])
            }

    monkeypatch.setattr(messagebox,
                        "showerror",
                        lambda *args,
                        **kwargs: None)
    monkeypatch.setattr("patient_gui.process_cpap_data",
                        mock_process_cpap_data)

    if expected_keys is None:
        metrics = process_cpap_data(file_path)
        assert metrics is None
    else:
        with pytest.raises(TypeError):
            process_cpap_data(file_path)


def test_display_metrics(monkeypatch):
    from patient_gui import display_metrics

    root = tk.Tk()
    metrics = {
        'breath_rate_bpm': 15,
        'apnea_count': 1
    }
    flow_rate = np.array([[0, 1], [1, 2], [2, 3]])

    # Mock the necessary functions and classes
    mock_label_config_called = False
    mock_subplots_called = False
    mock_savefig_called = False
    mock_open_called = False
    mock_photoimage_called = False
    mock_showerror_called = False

    class MockLabel:
        def __init__(self, *args, **kwargs):
            pass

        def grid(self, *args, **kwargs):
            pass

        def config(self, *args, **kwargs):
            nonlocal mock_label_config_called
            mock_label_config_called = True

    def mock_subplots(*args, **kwargs):
        nonlocal mock_subplots_called
        mock_subplots_called = True
        return plt.figure(), plt.axes()

    def mock_savefig(*args, **kwargs):
        nonlocal mock_savefig_called
        mock_savefig_called = True

    def mock_open(*args, **kwargs):
        nonlocal mock_open_called
        mock_open_called = True
        return Image.new('RGB', (100, 100))

    def mock_photoimage(*args, **kwargs):
        nonlocal mock_photoimage_called
        mock_photoimage_called = True

    def mock_showerror(*args, **kwargs):
        nonlocal mock_showerror_called
        mock_showerror_called = True

    monkeypatch.setattr(ttk, "Label", MockLabel)
    monkeypatch.setattr(plt, "subplots", mock_subplots)
    monkeypatch.setattr(plt, "savefig", mock_savefig)
    monkeypatch.setattr(Image, "open", mock_open)
    monkeypatch.setattr(ImageTk, "PhotoImage", mock_photoimage)
    monkeypatch.setattr(messagebox, "showerror", mock_showerror)

    breathing_rate_label = MockLabel()
    apnea_count_label = MockLabel()
    plot_label = MockLabel()

    display_metrics(root, metrics, flow_rate,
                    breathing_rate_label, apnea_count_label,
                    plot_label)

    # Assert that the necessary functions were called
    assert mock_label_config_called
    assert mock_subplots_called
    assert mock_open_called
    assert mock_photoimage_called
    assert not mock_showerror_called
